feat: sprint lifecycle fixes, Layer Protocol, WAL/IPC fixes, bounded_gather tuning
- Issue #5: Pandas 3.0 → Polars zero-copy path in duckdb_store
- Issue #6.1: Layer Protocol + LayerStack (UNIX domain socket zero-copy)
- Issue #10: WAL put_many per-item validation (list[bool] all() check)
- Issue #11: IPC batch memory fix (_build_single_arrow_batch extraction)
- Issue #17: asyncio.wrap_future pattern in parallel_scheduler
- Issue #22: snapshot race fix with threading.RLock in resource_governor
- Issue #8: Dual-channel TTL in resource_governor (io_only cache 5s→0.5s)
- P1-09: bounded_gather util (semaphore-gated gather) in async_helpers
- Coverage threshold fail_under=70 added to pyproject.toml
- 259 files changed, 6091 insertions, 436 deletions

+++ brain/ane_embedder.py
@@ -279,8 +279,9 @@ def _coreml_embed(model, text: str) -> np.ndarray:
         max_length=64,
         truncation=True,
     )
-    input_ids = tokens["input_ids"].astype(np.int32).flatten().tolist()
-    attn_mask  = tokens["attention_mask"].astype(np.int32).flatten().tolist()
+    # MLX/CoreML accept int64 tokenizer output directly — no .astype(np.int32) needed
+    input_ids = tokens["input_ids"].flatten().tolist()
+    attn_mask  = tokens["attention_mask"].flatten().tolist()
     feat_dict = {
         "input_ids":      _make_ml_array(input_ids),
         "attention_mask": _make_ml_array(attn_mask),
